Move tests of truncate_dataset to experimental
