+++ tests/test_data_utils.py
@@ -34,7 +34,6 @@ from trl.data_utils import (
     pack_dataset,
     prepare_multimodal_messages,
     prepare_multimodal_messages_vllm,
-    truncate_dataset,
     unpair_preference_dataset,
 )
 
@@ -1190,85 +1189,6 @@ class TestPackDatasetBfd(TrlTestCase):
             "seq_lengths": [[3, 1], [2]],
         }
         dataset = pack_dataset(dataset, seq_length, strategy="bfd_split")
-        assert dataset.to_dict() == expected_output
-
-
-class TestTruncateExamples(TrlTestCase):
-    def test_with_dataset(self):
-        examples = {
-            "input_ids": [[1, 2, 3], [4, 5, 6, 7], [8]],
-            "attention_mask": [[0, 1, 1], [0, 0, 1, 1], [1]],
-        }
-        dataset = Dataset.from_dict(examples)
-        dataset = dataset.with_format("numpy", dtype="float32")
-        format = dataset.format
-        max_length = 2
-        expected_output = {
-            "input_ids": [[1, 2], [4, 5], [8]],
-            "attention_mask": [[0, 1], [0, 0], [1]],
-        }
-        dataset = truncate_dataset(dataset, max_length)
-        assert dataset.to_dict() == expected_output
-        assert format == dataset.format
-
-    def test_with_iterable_dataset(self):
-        examples = {
-            "input_ids": [[1, 2, 3], [4, 5, 6, 7], [8]],
-            "attention_mask": [[0, 1, 1], [0, 0, 1, 1], [1]],
-        }
-        dataset = Dataset.from_dict(examples).to_iterable_dataset()
-        dataset = dataset.with_format("numpy")
-        formatting = dataset._formatting
-        max_length = 2
-        expected_output = {
-            "input_ids": [[1, 2], [4, 5], [8]],
-            "attention_mask": [[0, 1], [0, 0], [1]],
-        }
-        dataset = truncate_dataset(dataset, max_length)
-        num_examples = len(examples[next(iter(examples))])
-        assert next(iter(dataset.with_format(None).batch(batch_size=num_examples))) == expected_output
-        assert formatting == dataset._formatting
-
-    def test_with_extra_column(self):
-        examples = {
-            "input_ids": [[1, 2, 3], [4, 5, 6, 7], [8]],
-            "attention_mask": [[0, 1, 1], [0, 0, 1, 1], [1]],
-            "my_column": ["a", "b", "c"],
-        }
-        dataset = Dataset.from_dict(examples)
-        max_length = 2
-        expected_output = {
-            "input_ids": [[1, 2], [4, 5], [8]],
-            "attention_mask": [[0, 1], [0, 0], [1]],
-            "my_column": ["a", "b", "c"],
-        }
-        dataset = truncate_dataset(dataset, max_length)
-        assert dataset.to_dict() == expected_output
-
-    def test_with_keep_end(self):
-        examples = {
-            "input_ids": [[1, 2, 3], [4, 5, 6, 7], [8]],
-            "attention_mask": [[0, 1, 1], [0, 0, 1, 1], [1]],
-        }
-        dataset = Dataset.from_dict(examples)
-        expected_output = {
-            "input_ids": [[2, 3], [6, 7], [8]],
-            "attention_mask": [[1, 1], [1, 1], [1]],
-        }
-        dataset = truncate_dataset(dataset, max_length=2, truncation_mode="keep_end")
-        assert dataset.to_dict() == expected_output
-
-    def test_with_keep_end_and_zero_max_length(self):
-        examples = {
-            "input_ids": [[1, 2, 3], [4, 5, 6, 7], [8]],
-            "attention_mask": [[0, 1, 1], [0, 0, 1, 1], [1]],
-        }
-        dataset = Dataset.from_dict(examples)
-        expected_output = {
-            "input_ids": [[], [], []],
-            "attention_mask": [[], [], []],
-        }
-        dataset = truncate_dataset(dataset, max_length=0, truncation_mode="keep_end")
         assert dataset.to_dict() == expected_output
